# swagger/swagger.yaml
basePath: /api/v1
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  GoaContainerCreateResults:
    description: The results of container creation (default view)
    example:
      id: 1.794563140867684e+18
    properties:
      id:
        description: container id
        example: 1.794563140867684e+18
        format: int64
        type: integer
    required:
    - id
    title: 'Mediatype identifier: vnd.application/goa.container.create.results+json;
      view=default'
    type: object
  GoaContainerInspect:
    description: GoaContainerInspect media type (default view)
    example:
      args:
      - Aut rerum dolorem rerum aut nobis.
      created: "1994-12-21T10:17:26Z"
      id: 1.0549173781011334e+18
      image: Alias quas.
      imageID: Tenetur ut laudantium fugit aut officia inventore.
      name: Ex et nostrum quo aut.
      path: Ex totam et dolores quae sapiente.
      raw_state:
        dead: true
        exitCode: 9.010558457491618e+18
        finishedAt: "1972-07-23T17:34:47Z"
        oomKilled: false
        paused: false
        pid: 7.966749366891489e+17
        restarting: false
        running: false
        startedAt: "1976-08-09T23:40:57Z"
        status: running
      status: Created
      volumes:
      - Delectus libero non asperiores neque ut possimus.
      - Delectus libero non asperiores neque ut possimus.
      - Delectus libero non asperiores neque ut possimus.
    properties:
      args:
        description: The arguments to the command being run
        example:
        - Aut rerum dolorem rerum aut nobis.
        items:
          example: Aut rerum dolorem rerum aut nobis.
          type: string
        type: array
      created:
        description: The time the container was created
        example: "1994-12-21T10:17:26Z"
        format: date-time
        type: string
      id:
        description: ID
        example: 1.0549173781011334e+18
        format: int64
        type: integer
      image:
        description: The name of the image to use when creating the container
        example: Alias quas.
        type: string
      imageID:
        description: The container's image ID
        example: Tenetur ut laudantium fugit aut officia inventore.
        type: string
      name:
        description: Assign the specified name to the container. Must match /?[a-zA-Z0-9_-]+.
        example: Ex et nostrum quo aut.
        type: string
      path:
        description: The path to the command being run
        example: Ex totam et dolores quae sapiente.
        type: string
      raw_state:
        $ref: '#/definitions/GoaContainerInspectRaw_state'
      status:
        enum:
        - Image Downloading
        - Created
        - Running
        - Stopped
        - Error
        example: Created
        type: string
      volumes:
        description: Paths to mount volumes in
        example:
        - Delectus libero non asperiores neque ut possimus.
        - Delectus libero non asperiores neque ut possimus.
        - Delectus libero non asperiores neque ut possimus.
        items:
          example: Delectus libero non asperiores neque ut possimus.
          type: string
        type: array
    required:
    - name
    - id
    - image
    - imageID
    - path
    - args
    - created
    - status
    - raw_state
    - volumes
    title: 'Mediatype identifier: vpn.application/goa.container.inspect; view=default'
    type: object
  GoaContainerInspectRaw_state:
    description: GoaContainerInspectRaw_state media type (default view)
    example:
      dead: true
      exitCode: 9.010558457491618e+18
      finishedAt: "1972-07-23T17:34:47Z"
      oomKilled: false
      paused: false
      pid: 7.966749366891489e+17
      restarting: false
      running: false
      startedAt: "1976-08-09T23:40:57Z"
      status: running
    properties:
      dead:
        example: true
        type: boolean
      exitCode:
        example: 9.010558457491618e+18
        format: int64
        type: integer
      finishedAt:
        example: "1972-07-23T17:34:47Z"
        format: date-time
        type: string
      oomKilled:
        example: false
        type: boolean
      paused:
        example: false
        type: boolean
      pid:
        example: 7.966749366891489e+17
        format: int64
        type: integer
      restarting:
        example: false
        type: boolean
      running:
        example: false
        type: boolean
      startedAt:
        example: "1976-08-09T23:40:57Z"
        format: date-time
        type: string
      status:
        enum:
        - created
        - running
        - paused
        - restarting
        - removing
        - exited
        - dead
        example: running
        type: string
    required:
    - exitCode
    - finishedAt
    - oomKilled
    - dead
    - paused
    - pid
    - restarting
    - running
    - startedAt
    - status
    title: 'Mediatype identifier: vnd.application/goa.container.inspect.raw_state;
      view=default'
    type: object
...
  GoaContainerListEachCollection:
    description: GoaContainerListEachCollection is the media type for an array of
      GoaContainerListEach (default view)
    example:
    - command: Beatae culpa quia nisi dolore ut nisi.
      created: "2007-03-31T16:57:21Z"
      id: 4.875770818045676e+18
      image: Odio quia voluptate explicabo.
      imageID: Accusantium cumque nihil amet laborum suscipit.
      name: Consequatur aut neque quas cupiditate.
      status: Created
      volumes:
      - Et a dicta fugit qui quis.
    - command: Beatae culpa quia nisi dolore ut nisi.
      created: "2007-03-31T16:57:21Z"
      id: 4.875770818045676e+18
      image: Odio quia voluptate explicabo.
      imageID: Accusantium cumque nihil amet laborum suscipit.
      name: Consequatur aut neque quas cupiditate.
      status: Created
      volumes:
      - Et a dicta fugit qui quis.
    - command: Beatae culpa quia nisi dolore ut nisi.
      created: "2007-03-31T16:57:21Z"
      id: 4.875770818045676e+18
      image: Odio quia voluptate explicabo.
      imageID: Accusantium cumque nihil amet laborum suscipit.
      name: Consequatur aut neque quas cupiditate.
      status: Created
      volumes:
      - Et a dicta fugit qui quis.
    items:
      $ref: '#/definitions/GoaContainerListEach'
    title: 'Mediatype identifier: vpn.application/goa.container.list.each; type=collection;
      view=default'
    type: array
...
  SigninPayload:
    example:
      id: identify key
      password: ckqvt8597d
    properties:
      id:
        description: ID or Email
        example: identify key
        type: string
      password:
        description: Password
        example: ckqvt8597d
        maxLength: 256
        type: string
    required:
    - id
    - password
    title: SigninPayload
    type: object
...
  Vironsetting:
    description: viron setting (default view)
    example:
      color: Excepturi sapiente debitis quia alias ipsum.
      name: Et doloremque reiciendis ducimus minima labore odio.
      pages:
      - components:
        - api:
            method: Sunt facilis.
            path: Quam assumenda voluptatem autem nisi autem numquam.
          name: Et similique veniam odio.
          pagination: false
          primary: Reprehenderit quis.
          query:
          - key: Delectus fugiat qui nulla ipsa.
            type: Quo aut facere consequatur dolor.
          - key: Delectus fugiat qui nulla ipsa.
            type: Quo aut facere consequatur dolor.
          style: Qui tempore omnis quae aut quis blanditiis.
          table_labels:
          - Magni aut dolore similique.
          - Magni aut dolore similique.
          - Magni aut dolore similique.
        group: Quibusdam natus.
        id: Doloremque laudantium velit iure eum doloribus laudantium.
        name: Qui id sint et.
        section: Qui voluptatem quo.
      - components:
        - api:
            method: Sunt facilis.
            path: Quam assumenda voluptatem autem nisi autem numquam.
          name: Et similique veniam odio.
          pagination: false
          primary: Reprehenderit quis.
          query:
          - key: Delectus fugiat qui nulla ipsa.
            type: Quo aut facere consequatur dolor.
          - key: Delectus fugiat qui nulla ipsa.
            type: Quo aut facere consequatur dolor.
          style: Qui tempore omnis quae aut quis blanditiis.
          table_labels:
          - Magni aut dolore similique.
          - Magni aut dolore similique.
          - Magni aut dolore similique.
        group: Quibusdam natus.
        id: Doloremque laudantium velit iure eum doloribus laudantium.
        name: Qui id sint et.
        section: Qui voluptatem quo.
      tags:
      - Aut sunt minus aut quia omnis.
      - Aut sunt minus aut quia omnis.
      theme: Illum assumenda omnis tempora.
      thumbnail: Debitis non illo et ut et cumque.
    properties:
      color:
        description: color
        example: Excepturi sapiente debitis quia alias ipsum.
        type: string
      name:
        description: name
        example: Et doloremque reiciendis ducimus minima labore odio.
        type: string
      pages:
        description: pages
        example:
        - components:
          - api:
              method: Sunt facilis.
              path: Quam assumenda voluptatem autem nisi autem numquam.
            name: Et similique veniam odio.
            pagination: false
            primary: Reprehenderit quis.
            query:
            - key: Delectus fugiat qui nulla ipsa.
              type: Quo aut facere consequatur dolor.
            - key: Delectus fugiat qui nulla ipsa.
              type: Quo aut facere consequatur dolor.
            style: Qui tempore omnis quae aut quis blanditiis.
            table_labels:
            - Magni aut dolore similique.
            - Magni aut dolore similique.
            - Magni aut dolore similique.
          group: Quibusdam natus.
          id: Doloremque laudantium velit iure eum doloribus laudantium.
          name: Qui id sint et.
          section: Qui voluptatem quo.
        - components:
          - api:
              method: Sunt facilis.
              path: Quam assumenda voluptatem autem nisi autem numquam.
            name: Et similique veniam odio.
            pagination: false
            primary: Reprehenderit quis.
            query:
            - key: Delectus fugiat qui nulla ipsa.
              type: Quo aut facere consequatur dolor.
            - key: Delectus fugiat qui nulla ipsa.
              type: Quo aut facere consequatur dolor.
            style: Qui tempore omnis quae aut quis blanditiis.
            table_labels:
            - Magni aut dolore similique.
            - Magni aut dolore similique.
            - Magni aut dolore similique.
          group: Quibusdam natus.
          id: Doloremque laudantium velit iure eum doloribus laudantium.
          name: Qui id sint et.
          section: Qui voluptatem quo.
        items:
          $ref: '#/definitions/Vironpage'
        type: array
      tags:
        description: tags
        example:
        - Aut sunt minus aut quia omnis.
        - Aut sunt minus aut quia omnis.
        items:
          example: Aut sunt minus aut quia omnis.
          type: string
        type: array
      theme:
        description: theme
        example: Illum assumenda omnis tempora.
        type: string
      thumbnail:
        description: thumbnail
        example: Debitis non illo et ut et cumque.
        type: string
    required:
    - name
    - color
    - theme
    - pages
    - tags
    - thumbnail
    title: 'Mediatype identifier: application/vnd.vironsetting+json; view=default'
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609066e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609066e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
host: localhost:4434
info:
  title: Modoki API Documentation
  version: ""
paths:
  /container/create:
    get:
      description: create a new container
      operationId: container#create
      parameters:
      - collectionFormat: multi
        description: Command to run specified as a string or an array of strings.
        in: query
        items:
          type: string
        name: command
        required: false
        type: array
      - collectionFormat: multi
        description: The entry point for the container as a string or an array of
          strings
        in: query
        items:
          type: string
        name: entrypoint
        required: false
        type: array
      - collectionFormat: multi
        description: Environment variables
        in: query
        items:
          type: string
        name: env
        required: false
        type: array
      - description: Name of image
        in: query
        name: image
        required: true
        type: string
      - description: Name of container and subdomain
        in: query
        maxLength: 64
        name: name
        pattern: ^[a-zA-Z0-9_]+$
        required: true
        type: string
      - default: true
        description: Whether HTTP is redirected to HTTPS
        in: query
        name: sslRedirect
        required: false
        type: boolean
      - collectionFormat: multi
        description: Path to volumes in a container
        in: query
        items:
          type: string
        name: volumes
        required: false
        type: array
      - description: Current directory (PWD) in the command will be launched
        in: query
        name: workingDir
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - vnd.application/goa.container.create.results+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaContainerCreateResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: create container
      tags:
      - container
  /container/download:
    get:
      description: Copy files from the container
      operationId: container#download
      parameters:
      - description: ID or name
        in: query
        name: id
        required: true
        type: string
      - description: Path in the container to save files
        in: query
        name: internalPath
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: download container
      tags:
      - container
  /container/inspect:
    get:
      description: Return details of a container
      operationId: container#inspect
      parameters:
      - description: ID or name
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - vpn.application/goa.container.inspect
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaContainerInspect'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: inspect container
      tags:
      - container
  /container/list:
    get:
      description: Return a list of containers
      operationId: container#list
      produces:
      - application/vnd.goa.error
      - vpn.application/goa.container.list.each; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaContainerListEachCollection'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: list container
      tags:
      - container
  /container/logs:
    get:
      description: Get stdout and stderr logs from a container.
      operationId: container#logs
      parameters:
      - default: false
        in: query
        name: follow
        required: false
        type: boolean
      - description: id or name
        in: query
        name: id
        required: true
        type: string
      - in: query
        name: since
        required: false
        type: string
      - default: false
        in: query
        name: stderr
        required: false
        type: boolean
      - default: false
        in: query
        name: stdout
        required: false
        type: boolean
      - default: all
        in: query
        name: tail
        required: false
        type: string
      - default: false
        in: query
        name: timestamps
        required: false
        type: boolean
      - in: query
        name: until
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      responses:
        "101":
          description: Switching Protocols
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - ws
      security:
      - jwt: []
      summary: logs container
      tags:
      - container
  /container/remove:
    get:
      description: remove a container
      operationId: container#remove
      parameters:
      - default: false
        description: If the container is running, kill it before removing it.
        in: query
        name: force
        required: true
        type: boolean
      - description: id or name
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "409":
          description: You cannot remove a running container
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: remove container
      tags:
      - container
  /container/start:
    get:
      description: start a container
      operationId: container#start
      parameters:
      - description: id or name
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: start container
      tags:
      - container
  /container/stop:
    get:
      description: stop a container
      operationId: container#stop
      parameters:
      - description: id or name
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: stop container
      tags:
      - container
  /container/upload:
    post:
      description: Copy files to the container
      operationId: container#upload
      parameters:
      - default: false
        description: Allow for a existing directory to be replaced by a file
        in: formData
        name: allowOverwrite
        required: false
        type: boolean
      - description: File tar archive
        in: formData
        name: data
        required: true
        type: file
      - description: ID or name
        in: formData
        name: id
        required: true
        type: string
      - description: Path in the container to save files
        in: formData
        name: path
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "413":
          description: Request Entity Too Large
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: upload container
      tags:
      - container
  /signin:
    post:
      description: Creates a valid JWT
      operationId: viron#signin
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/SigninPayload'
      responses:
        "204":
          description: No Content
          headers:
            Authorization:
              description: Generated JWT
              type: string
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      schemes:
      - http
      - https
      summary: signin viron
      tags:
      - viron
  /viron:
    get:
      description: Get viron menu
      operationId: viron#get
      produces:
      - application/vnd.vironsetting+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Vironsetting'
      schemes:
      - http
      - https
      security:
      - jwt: []
      summary: get viron
      tags:
      - viron
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- http
- https
securityDefinitions:
  jwt:
    description: |2-


      **Security Scopes**:
        * `api:access`: API access
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"